% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioDataScience2-package.R
\docType{package}
\name{BioDataScience2-package}
\alias{BioDataScience2}
\alias{BioDataScience2-package}
\title{A series of learnr tutorials and Shiny applications for BioDataScience2}
\description{
Run the tutorials using\code{\link[=run]{run()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioDataScience-Course/BioDataScience2}
  \item Report bugs at \url{https://github.com/BioDataScience-Course/BioDataScience2/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org}

Authors:
\itemize{
  \item Guyliann Engels \email{guyliann.engels@umons.ac.be}
}

}
