% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioDataScience2-package.R
\docType{package}
\name{BioDataScience2-package}
\alias{BioDataScience2-package}
\title{A series of learnr tutorials and Shiny applications for BioDataScience2}
\description{
Run the tutorials using\code{\link[=run]{run()}}.
}
